feat: Implement latest manga page
